Move wall clock-based time position into Spotify backend
