+++ mopidy/backends/spotify/playback.py
@@ -1,4 +1,5 @@
 import logging
+import time
 
 from spotify import Link, SpotifyError
 
@@ -10,11 +11,25 @@ logger = logging.getLogger('mopidy.backends.spotify.playback')
 
 
 class SpotifyPlaybackProvider(BasePlaybackProvider):
+    def __init__(self, *args, **kwargs):
+        super(SpotifyPlaybackProvider, self).__init__(*args, **kwargs)
+
+        self._play_time_accumulated = 0
+        self._play_time_started = 0
+
+    def pause(self):
+        time_since_started = self._wall_time() - self._play_time_started
+        self._play_time_accumulated += time_since_started
+
+        return super(SpotifyPlaybackProvider, self).pause()
+
     def play(self, track):
-        if self.backend.playback.state == PlaybackState.PLAYING:
-            self.backend.spotify.session.play(0)
         if track.uri is None:
             return False
+
+        self._play_time_accumulated = 0
+        self._play_time_started = self._wall_time()
+
         try:
             self.backend.spotify.session.load(
                 Link.from_string(track.uri).as_track())
@@ -29,12 +44,17 @@ class SpotifyPlaybackProvider(BasePlaybackProvider):
             return False
 
     def resume(self):
+        self._play_time_started = self._wall_time()
         return self.seek(self.backend.playback.time_position)
 
     def seek(self, time_position):
+        self._play_time_started = self._wall_time()
+        self._play_time_accumulated = time_position
+
         self.backend.audio.prepare_change()
         self.backend.spotify.session.seek(time_position)
         self.backend.audio.start_playback()
+
         return True
 
     def stop(self):
@@ -46,4 +66,15 @@ class SpotifyPlaybackProvider(BasePlaybackProvider):
         # when used with the Spotify backend and GStreamer appsrc. If this can
         # be resolved, we no longer need to use a wall clock based time
         # position for Spotify playback.
-        return self.wall_clock_based_time_position()
+        state = self.backend.playback.state
+        if state == PlaybackState.PLAYING:
+            time_since_started = (self._wall_time() -
+                self._play_time_started)
+            return self._play_time_accumulated + time_since_started
+        elif state == PlaybackState.PAUSED:
+            return self._play_time_accumulated
+        elif state == PlaybackState.STOPPED:
+            return 0
+
+    def _wall_time(self):
+        return int(time.time() * 1000)
